Move rotation matrix code to Numerics
Also make fix in CUDA code.
